ss_extract: Option to extract duets to single file
This creates a single duet-style txt file instead of separate files for
each singer. In addition, the program now finds duets that have a single
track outside of the TRACK tags. In that case, the singer for any given
note is determined by the last seen Singer attribute in a SENTENCE tag.