Extract medley start and end beats
The medley start and end time in seconds are extracted from the
songs_*.xml files and then converted to beats. Medley start and end are
also marked at the corresponding notes in the melody files, but these
might be ambiguous in the case of duets. The "normal" medleys are always
chosen instead of the "micro" medleys.